requirements.txt
python-dotenv==*
asyncpg==*
discord.py==*
